test(rules): renamed package
