--- library/src/test/java/com/matthewtamlin/avatar/rules/avatar_rule/after_running_rule/Annotation1.java
@@ -1,3 +1,0 @@
-package com.matthewtamlin.avatar.rules.avatar_rule.after_running_rule;
-
-public @interface Annotation1 {}
