+++ library/src/test/java/com/matthewtamlin/avatar/rules/avatar_rule/running_regular_rule/Annotation1.java
@@ -1,0 +1,3 @@
+package com.matthewtamlin.avatar.rules.avatar_rule.running_regular_rule;
+
+public @interface Annotation1 {}
